asfdemux: More index handling
In the same vein as the previous commit, the end of the file might have been
cut and therefore contains bogus data.

If the object is too short, consider it as garbage and return a non-fatal
flow.

Regression introduced in 97294eb8bbed1b9dad7d3f2c52dd69eb1812cc06